+++ src/IBM.WatsonDeveloperCloud.LanguageTranslator/v2/LanguageTranslatorService.cs
@@ -22,7 +22,7 @@ using System.Net.Http;
 using System.Net.Http.Headers;
 using System.Text;
 using IBM.WatsonDeveloperCloud.Http;
-using IBM.WatsonDeveloperCloud.LanguageTranslator.v2.Models;
+using IBM.WatsonDeveloperCloud.LanguageTranslator.v2.Model;
 using IBM.WatsonDeveloperCloud.Service;
 using Newtonsoft.Json.Linq;
 using System.Runtime.ExceptionServices;
@@ -85,7 +85,7 @@ namespace IBM.WatsonDeveloperCloud.LanguageTranslator.v2
             }
             catch (AggregateException ae)
             {
-                ExceptionDispatchInfo.Capture(ae.Flatten().InnerException).Throw(); ;
+                throw ae.Flatten();
             }
 
             return result;
@@ -120,7 +120,7 @@ namespace IBM.WatsonDeveloperCloud.LanguageTranslator.v2
             }
             catch (AggregateException ae)
             {
-                ExceptionDispatchInfo.Capture(ae.Flatten().InnerException).Throw(); ;
+                throw ae.Flatten();
             }
 
             return result;
@@ -143,7 +143,7 @@ namespace IBM.WatsonDeveloperCloud.LanguageTranslator.v2
             }
             catch (AggregateException ae)
             {
-                ExceptionDispatchInfo.Capture(ae.Flatten().InnerException).Throw(); ;
+                throw ae.Flatten();
             }
 
             return result;
@@ -166,7 +166,7 @@ namespace IBM.WatsonDeveloperCloud.LanguageTranslator.v2
             }
             catch (AggregateException ae)
             {
-                ExceptionDispatchInfo.Capture(ae.Flatten().InnerException).Throw(); ;
+                throw ae.Flatten();
             }
 
             return result;
@@ -221,7 +221,7 @@ namespace IBM.WatsonDeveloperCloud.LanguageTranslator.v2
             }
             catch (AggregateException ae)
             {
-                ExceptionDispatchInfo.Capture(ae.Flatten().InnerException).Throw(); ;
+                throw ae.Flatten();
             }
 
             return result;
@@ -241,7 +241,7 @@ namespace IBM.WatsonDeveloperCloud.LanguageTranslator.v2
             }
             catch (AggregateException ae)
             {
-                ExceptionDispatchInfo.Capture(ae.Flatten().InnerException).Throw(); ;
+                throw ae.Flatten();
             }
 
             return result;
@@ -263,7 +263,7 @@ namespace IBM.WatsonDeveloperCloud.LanguageTranslator.v2
             }
             catch (AggregateException ae)
             {
-                ExceptionDispatchInfo.Capture(ae.Flatten().InnerException).Throw(); ;
+                throw ae.Flatten();
             }
 
             return result;
